extract the queue part of the index-scheduler in another module
